+++ src/shared/types/index.ts
@@ -1,2 +1,2 @@
-export { ClassConstructor } from "./class-constructor.type";
-export { TimeMarker } from "./time-marker.type";
+export type { ClassConstructor } from "./class-constructor.type";
+export type { TimeMarker } from "./time-marker.type";